Address review comments and refactor code
